+++ example/lib/src/module/popup_page.dart
@@ -8,109 +8,78 @@ class PopupPage extends StatelessWidget {
   const PopupPage({super.key});
 
   @override
-  Widget build(BuildContext context) => ExtendedScaffold(
-          padding: const EdgeInsets.symmetric(horizontal: 20),
-          appBar: AppBarText('Popup'),
-          isScroll: true,
-          children: [
-            const Partition('Dialog'),
-            ElevatedText('popup Dialog', onTap: () {
-              const Center(child: _Dialog()).popupDialog<dynamic>();
-            }),
-            ElevatedText('popup MaterialDialog', onTap: () {
-              const Center(child: _Dialog()).popupMaterialDialog<dynamic>();
-            }),
-            ElevatedText('popup CupertinoDialog', onTap: () {
-              const Center(child: _Dialog()).popupCupertinoDialog<dynamic>();
-            }),
-            const Partition('Bottom Sheet'),
-            ElevatedText('popup BottomSheet', onTap: () {
-              const _ActionSheet().popupBottomSheet<dynamic>(
-                  options: const BottomSheetOptions(
-                      backgroundColor: Colors.transparent));
-            }),
-            ElevatedText('popup BottomSheet - Full screen', onTap: () {
-              Container(
-                      color: Colors.blueGrey.withOpacity(0.4),
-                      alignment: Alignment.center,
-                      child: const _ActionSheet())
-                  .popupBottomSheet<dynamic>(
-                      options: const BottomSheetOptions(
-                          backgroundColor: Colors.transparent));
-            }),
-            ElevatedText('popup CupertinoModal', onTap: () {
-              const _ActionSheet().popupCupertinoModal<dynamic>();
-            }),
-            const Partition('DoubleChooseWindows Popup'),
-            ElevatedText('popup DoubleChooseWindows',
-                onTap: () => doubleChooseWindows(context)),
-          ]);
-
-  void doubleChooseWindows(BuildContext context) {
-    DoubleChooseWindows(
-            options: const ModalBoxOptions(),
-            decoration: BoxDecoration(
-                color: context.theme.scaffoldBackgroundColor,
-                borderRadius: BorderRadius.circular(10)),
-            right: Universal(
-                padding: const EdgeInsets.symmetric(vertical: 6),
-                alignment: Alignment.center,
-                child: Text('确定', style: context.textTheme.titleMedium),
-                onTap: () {
-                  ///如果isOverlay=true; 必须先closeOverlay() 再toast或者loading
-                  showToast('确定');
-                }),
-            left: Universal(
-                alignment: Alignment.center,
-                padding: const EdgeInsets.symmetric(vertical: 6),
-                child: Text('取消', style: context.textTheme.titleMedium),
-                onTap: () {
-                  ///如果isOverlay=true; 必须先closeOverlay() 再toast或者loading
-                  showToast('取消');
-                  pop();
-                }),
-            content: Padding(
-                padding: const EdgeInsets.symmetric(vertical: 40),
-                child: Text('内容', style: context.textTheme.bodyLarge)))
-        .show();
-  }
-}
-
-class _Dialog extends StatelessWidget {
-  const _Dialog();
-
-  @override
   Widget build(BuildContext context) {
-    return DoubleChooseWindows(
-        decoration: BoxDecoration(
-            color: context.theme.dialogBackgroundColor,
+    final actionsDialog = ActionDialog(
+        options: ModalBoxOptions(
+            foregroundColor: context.theme.dialogBackgroundColor,
             borderRadius: BorderRadius.circular(10)),
-        content: Container(
-            margin: const EdgeInsets.all(12),
+        dividerColor: const Color(0xFF8F8F8F),
+        title: const Universal(
+            alignment: Alignment.center, height: 50, child: Text('Title')),
+        content: Universal(
+            margin: const EdgeInsets.all(10),
             alignment: Alignment.center,
             decoration: BoxDecoration(
                 color: Colors.grey.withOpacity(0.4),
                 borderRadius: BorderRadius.circular(4)),
             height: 200,
-            child: const Text('Dialog')),
-        left: Universal(
-            margin: const EdgeInsets.all(4),
-            alignment: Alignment.center,
-            onTap: () {
-              showToast('left');
-              pop();
-            },
-            unifiedButtonCategory: UnifiedButtonCategory.outlined,
-            child: const Text('left')),
-        right: Universal(
-            margin: const EdgeInsets.all(4),
-            alignment: Alignment.center,
-            onTap: () {
-              showToast('right');
-              pop();
-            },
-            unifiedButtonCategory: UnifiedButtonCategory.filled,
-            child: const Text('right')));
+            child: const Text('ActionDialog')),
+        actions: [
+          Universal(
+              expanded: true,
+              alignment: Alignment.center,
+              onTap: () {
+                showToast('Left');
+                pop();
+              },
+              child: const Text('Left')),
+          Universal(
+              expanded: true,
+              alignment: Alignment.center,
+              onTap: () {
+                showToast('Center');
+                pop();
+              },
+              child: const Text('Center')),
+          Universal(
+              expanded: true,
+              alignment: Alignment.center,
+              onTap: () {
+                showToast('Right');
+                pop();
+              },
+              child: const Text('Right'))
+        ]);
+    return ExtendedScaffold(
+        padding: const EdgeInsets.symmetric(horizontal: 20),
+        appBar: AppBarText('Popup'),
+        isScroll: true,
+        children: [
+          const Partition('ActionDialog'),
+          ElevatedText('popup Dialog', onTap: () {
+            actionsDialog.popupDialog<dynamic>();
+          }),
+          ElevatedText('popup MaterialDialog', onTap: () {
+            actionsDialog.popupMaterialDialog<dynamic>();
+          }),
+          ElevatedText('popup CupertinoDialog', onTap: () {
+            actionsDialog.popupCupertinoDialog<dynamic>();
+          }),
+          const Partition('Bottom Sheet'),
+          ElevatedText('popup BottomSheet', onTap: () {
+            const _ActionSheet().popupBottomSheet<dynamic>(
+                options: const BottomSheetOptions(
+                    backgroundColor: Colors.transparent));
+          }),
+          ElevatedText('popup BottomSheet - Full screen', onTap: () {
+            const _ActionSheet().popupBottomSheet<dynamic>(
+                options: const BottomSheetOptions(
+                    backgroundColor: Colors.transparent));
+          }),
+          ElevatedText('popup CupertinoModal', onTap: () {
+            const _ActionSheet().popupCupertinoModal<dynamic>();
+          }),
+        ]);
   }
 }