+++ lib/src/extension/object_extension.dart
@@ -38,48 +38,6 @@ extension ExtensionT<T> on T {
   /// 转为 ValueNotifier
   ValueNotifier<T> get notifier => ValueNotifier<T>(this);
 
-  /// toast 显示
-  Future<ExtendedOverlayEntry?> showToast({
-    /// icon 样式 不传 仅显示文字
-    ToastIconStyle? iconStyle,
-
-    /// 动画样式
-    FlAnimationStyle? animationStyle,
-
-    /// 动画执行时间
-    Duration? animationDuration,
-
-    /// 显示文字样式
-    TextStyle? textStyle,
-
-    /// Toast显示时间
-    Duration? duration,
-
-    /// Toast 装饰器 会替换 [backgroundColor]
-    BoxDecoration? decoration,
-
-    /// Toast onTap
-    GestureTapCallback? onToastTap,
-
-    /// builder
-    ToastBuilder? builder,
-
-    /// ToastOptions
-    ToastOptions? options,
-  }) =>
-      Toast(
-        toString(),
-        builder: builder,
-        onToastTap: onToastTap,
-        decoration: decoration,
-        duration: duration,
-        animationStyle: animationStyle,
-        animationDuration: animationDuration,
-        textStyle: textStyle,
-        iconStyle: iconStyle,
-        options: options,
-      ).show();
-
   List<T> get toList => [this];
 
   void log({bool? crossLine}) {